finish all code

+++ src/Thingsboard.Net.Abstractions/TbQueueModels/TbQueueProcessingStrategy.cs
@@ -2,7 +2,7 @@
 
 public class TbQueueProcessingStrategy
 {
-    public int     FailurePercentage      { get; set; }
+    public double  FailurePercentage      { get; set; }
     public int     MaxPauseBetweenRetries { get; set; }
     public int     PauseBetweenRetries    { get; set; }
     public int     Retries                { get; set; }
